feat: add aggregate data

+++ lib/features/usage_monitor/domain/entity/usage_monitor_entity.dart
@@ -1,8 +1,8 @@
 import 'package:equatable/equatable.dart';
 
 class UsageMonitorEntity extends Equatable {
-  final int timestamp;
-  final int value;
+  final double timestamp;
+  final double value;
 
   const UsageMonitorEntity(this.timestamp, this.value);
